requirements.txt
apache-airflow==1.10.5
Flask-Bcrypt==0.7.1
numpy==1.17.0
pandas==0.25.0
pmdarima==1.3.0
seaborn==0.9.0
statsmodels==0.10.1